WIP Slices CSS refactor
