+++ site/pages/our-work/slices/car-trawler/index.js
@@ -1,3 +1,5 @@
+// @flow
+
 import React from 'react';
 
 import styles from '../shared/styles.css';
