+++ site/pages/our-work/slices/fortnum-and-mason-digital-transformation/index.js
@@ -1,12 +1,15 @@
+// @flow
+
 import React from 'react';
-import styles from './style.css';
+import sharedStyles from '../shared/styles.css';
+import styles from './styles.css';
 import Link from '../../../../components/link';
 
 import fmLogo from './images/fortnum-mason-logo.png';
 
 export default function CaseStudyFortnumDigitalTransformationSlice() {
   return (
-    <div className={styles.caseStudyContainer}>
+    <div className={sharedStyles.caseStudyContainer}>
       <div className={styles.caseStudyContent}>
         <div className={styles.caseStudyTextContainer}>
           <Link to="fMTeaCaseStudy">
@@ -14,11 +17,11 @@ export default function CaseStudyFortnumDigitalTransformationSlice() {
             <h2 className={styles.caseStudyTextContainerHeader}>
               The ongoing digital transformation of a 310-year old retailer
             </h2>
-            <p className={styles.description}>
+            <p className={sharedStyles.description}>
               Discover how we helped customers find products faster and drive sales of tea.
             </p>
             <div className={styles.links}>
-              <p className={styles.readmore}>Read more</p>
+              <p className={sharedStyles.readmore}>Read more</p>
             </div>
           </Link>
         </div>
