--- site/pages/our-work/slices/fortnum-and-mason/style.css
@@ -1,175 +1,0 @@
-@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
-@value mediumScreen, largeScreen from "../../../../css/_sizes.css";
-
-.caseStudyContainer {
-  max-width: 1100px;
-  justify-content: center;
-  margin: 0 auto;
-}
-
-.caseStudyContent {
-  display: flex;
-  max-width: containerMaxWidth;
-  width: 100%;
-  flex-direction: column-reverse;
-  padding-bottom: 20px;
-}
-
-.imageWrapper {
-  display: flex;
-  flex:1;
-  justify-content: center;
-  align-items: flex-end;
-}
-
-_:-ms-fullscreen, :root .imageWrapper {
-  min-height: 20vh;
-}
-
-.imageWrapper img {
-  min-width: 260px;
-  width: 70%;
-}
-
-.clientLogo {
-  width: 180px;
-}
-
-.caseStudyTextContainer {
-  padding: 0;
-  margin-top: 30px;
-}
-
-.caseStudyTextContainerHeader {
-  composes: fontM from "../../../../css/typography/_fonts.css";
-  composes: serif from "../../../../css/typography/_fonts.css";
-  color: badgerBlack;
-  margin-top: 30px;
-}
-
-.description {
-  composes: fontXS from "../../../../css/typography/_fonts.css";
-  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
-  color: badgerBlack;
-  margin-top: 10px;
-}
-
-.readmore {
-  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
-  border-bottom: 2px solid badgerRedOnWhite;
-  text-decoration: none;
-  color: badgerBlack;
-  display: inline-block;
-  margin-top: 30px;
-}
-
-.readmore:hover {
-  color: badgerRedOnWhite;
-}
-
-.links {
-  text-align: right;
-  border-bottom: 1px solid linesOnWhite;
-  padding-bottom: 30px;
-}
-
-.projectBigSmallSnapshot {
-  display: block;
-}
-
-.projectMediumSnapshot {
-  display: none;
-}
-
-@media mediumScreen {
-  .caseStudyContainer {
-    padding-bottom: 60px;
-    border-bottom: 1px solid linesOnWhite;
-  }
-
-  .caseStudyTextContainer {
-    margin-top: 0;
-    padding: 0 30px;
-    width: 60%;
-    position: relative;
-    z-index: 1;
-  }
-
-  .caseStudyContent {
-    display: block;
-    margin: 0;
-    padding: 0;
-    position: relative;
-  }
-
-  .imageWrapper {
-    position: absolute;
-    right: 0;
-    bottom: 40px;
-  }
-
-  .imageWrapper img {
-    width: 300px;
-  }
-
-  .clientLogo {
-    width: 230px;
-  }
-
-  .caseStudyTextContainerHeader {
-    margin-top: 40px;
-  }
-
-  .description {
-    margin-top: 20px;
-    max-width: 300px;
-  }
-
-  .caseStudyContainer {
-    padding-top: 40px;
-  }
-
-  .links {
-    text-align: left;
-    border: none;
-    padding-bottom: 0;
-  }
-
-  .projectBigSmallSnapshot {
-    display: none;
-  }
-
-  .projectMediumSnapshot {
-    display: block;
-  }
-}
-
-@media screen and (min-width: 800px) {
-  .imageWrapper img {
-    width: 400px;
-  }
-}
-
-@media screen and (min-width: 1150px) {
-  .projectBigSmallSnapshot {
-    display: block;
-  }
-
-  .projectMediumSnapshot {
-    display: none;
-  }
-
-  .imageWrapper {
-    position: absolute;
-    right: 0;
-    top: 50px;
-  }
-
-  .imageWrapper img {
-    width: 800px;
-  }
-
-  .caseStudyTextContainer {
-    max-width: 500px;
-  }
-}
