+++ site/pages/our-work/slices/shared/index.js
@@ -10,10 +10,11 @@ type Props = {
   children: React.ChildrenArray<React.Element<'div'>>,
   layout?: 'row' | 'row-reverse',
   to: string,
+  flushBottom?: boolean,
 };
 
-export const SliceContainer = ({ children, layout = 'row', to }: Props) => (
-  <div className={styles.caseStudyContainer}>
+export const SliceContainer = ({ children, layout = 'row', flushBottom = false, to }: Props) => (
+  <div className={flushBottom ? styles.caseStudyContainerFlushBottom : styles.caseStudyContainer}>
     <Link to={to} className={classNames(styles.caseStudyContent, styles[layout])}>
       {children}
     </Link>
