+++ site/pages/our-work/slices/sky/index.js
@@ -1,12 +1,15 @@
+// @flow
+
 import React from 'react';
-import styles from './style.css';
+import sharedStyles from '../shared/styles.css';
+import styles from './styles.css';
 import Link from '../../../../components/link';
 
 import skyLogo from './images/sky.png';
 
 export default function CaseStudySkySlice() {
   return (
-    <div className={styles.caseStudyContainer}>
+    <div className={sharedStyles.caseStudyContainer}>
       <div className={styles.caseStudyContent}>
         <div className={styles.caseStudyTextContainer}>
           <Link to="skyCMSCaseStudy">
@@ -14,11 +17,11 @@ export default function CaseStudySkySlice() {
             <h2 className={styles.caseStudyTextContainerHeader}>
               A modern CMS that is the foundation that supports both customers and the internal team
             </h2>
-            <p className={styles.description}>
+            <p className={sharedStyles.description}>
               Enabling Sky to manage their content with an amazing customer experience
             </p>
             <div className={styles.links}>
-              <p className={styles.readmore}>Read more</p>
+              <p className={sharedStyles.readmore}>Read more</p>
             </div>
           </Link>
         </div>
