--- site/pages/our-work/slices/sky/style.css
@@ -1,95 +1,0 @@
-@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
-@value mediumScreen, largeScreen from "../../../../css/_sizes.css";
-
-.caseStudyContainer {
-  max-width: 1100px;
-  justify-content: center;
-  margin: 0 auto;
-}
-
-.caseStudyContent {
-  display: flex;
-  max-width: containerMaxWidth;
-  width: 100%;
-  flex-direction: column-reverse;
-  margin-bottom: 20px;
-  padding-bottom: 20px;
-}
-
-.clientLogo {
-  width: 40px;
-}
-
-.caseStudyTextContainer {
-  padding: 0;
-}
-
-.caseStudyTextContainerHeader {
-  composes: fontL from "../../../../css/typography/_fonts.css";
-  composes: serif from "../../../../css/typography/_fonts.css";
-  color: badgerBlack;
-  margin-top: 30px;
-}
-
-.description {
-  composes: fontXS from "../../../../css/typography/_fonts.css";
-  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
-  color: badgerBlack;
-  margin-top: 10px;
-}
-
-.readmore {
-  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
-  border-bottom: 2px solid badgerRedOnWhite;
-  text-decoration: none;
-  color: badgerBlack;
-  display: inline-block;
-  margin-top: 30px;
-}
-
-.readmore:hover {
-  color: badgerRedOnWhite;
-}
-
-.links {
-  text-align: right;
-  border-bottom: 1px solid linesOnWhite;
-  padding-bottom: 30px;
-}
-
-@media mediumScreen {
-  .caseStudyTextContainer {
-    margin-top: 0;
-    padding: 0 30px;
-    text-align: center;
-  }
-
-  .caseStudyContent {
-    flex-direction: row;
-    margin: 0;
-    padding: 0;
-    border-bottom: 1px solid linesOnWhite;
-  }
-
-  .clientLogo {
-    width: 70px;
-  }
-
-  .caseStudyTextContainerHeader {
-    margin-top: 20px;
-  }
-
-  .description {
-    margin-top: 20px;
-  }
-
-  .caseStudyContainer {
-    padding-top: 40px;
-  }
-
-  .links {
-    text-align: center;
-    border: none;
-    padding-bottom: 60px;
-  }
-}
