+++ site/pages/our-work/style.css
@@ -12,25 +12,26 @@
 }
 
 .mainHeader {
-  composes: fontL from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontL from '../../css/typography/_fonts.css';
+  composes: serif from '../../css/typography/_fonts.css';
   color: badgerWhite;
 }
 
 .mainSlogan {
-  composes: fontM2 from "../../css/typography/_fonts.css";
-  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontM2 from '../../css/typography/_fonts.css';
+  composes: serif from '../../css/typography/_fonts.css';
   text-align: left;
   color: badgerWhite;
 }
 
 .caseStudyTopSection {
-  composes: dividerWhite from "../../css/_divider.css";
+  composes: dividerWhite from '../../css/_divider.css';
   background-color: badgerWhite;
 }
 
 .topSlicesContainer {
   margin: 0 5%;
+  padding-top: 30px;
 }
 
 .grid {
@@ -71,7 +72,7 @@
   }
 
   .mainHeader {
-    composes: fontL from "../../css/typography/_fonts.css";
+    composes: fontL from '../../css/typography/_fonts.css';
     margin-bottom: 10px;
   }
